@@ -30,20 +30,11 @@ namespace Corvus.CosmosClient.Extensions
         /// <returns>A <see cref="Task"/> which provides a continuation token if it terminates before .</returns>
         public static Task<string?> ForEachAsync<T>(this Container container, string queryText, Action<T> action, QueryRequestOptions? requestOptions = null, int? maxBatchCount = null, string? continuationToken = null, CancellationToken cancellationToken = default)
         {
-            if (container is null)
-            {
-                throw new ArgumentNullException(nameof(container));
-            }
+            ArgumentNullException.ThrowIfNull(container);
 
-            if (queryText is null)
-            {
-                throw new ArgumentNullException(nameof(queryText));
-            }
+            ArgumentNullException.ThrowIfNull(queryText);
 
-            if (action is null)
-            {
-                throw new ArgumentNullException(nameof(action));
-            }
+            ArgumentNullException.ThrowIfNull(action);
 
             return ForEachAsync(container, new QueryDefinition(queryText), action, requestOptions, maxBatchCount, continuationToken, cancellationToken);
         }
@@ -62,20 +53,11 @@ namespace Corvus.CosmosClient.Extensions
         /// <returns>A <see cref="Task"/> which provides a continuation token if it terminates before .</returns>
         public static async Task<string?> ForEachAsync<T>(this Container container, QueryDefinition queryDefinition, Action<T> action, QueryRequestOptions? requestOptions = null, int? maxBatchCount = null, string? continuationToken = null, CancellationToken cancellationToken = default)
         {
-            if (container is null)
-            {
-                throw new ArgumentNullException(nameof(container));
-            }
+            ArgumentNullException.ThrowIfNull(container);
 
-            if (queryDefinition == null)
-            {
-                throw new ArgumentNullException(nameof(queryDefinition));
-            }
+            ArgumentNullException.ThrowIfNull(queryDefinition);
 
-            if (action == null)
-            {
-                throw new ArgumentNullException(nameof(action));
-            }
+            ArgumentNullException.ThrowIfNull(action);
 
             FeedIterator<T> iterator = container.GetItemQueryIterator<T>(queryDefinition, continuationToken, requestOptions);
 
@@ -133,20 +115,11 @@ namespace Corvus.CosmosClient.Extensions
         /// <returns>A <see cref="Task"/> which provides a continuation token if it terminates before .</returns>
         public static Task<string?> ForEachAsync<T>(this Container container, string queryText, Func<T, Task> actionAsync, QueryRequestOptions? requestOptions = null, int? maxBatchCount = null, string? continuationToken = null, CancellationToken cancellationToken = default)
         {
-            if (container is null)
-            {
-                throw new ArgumentNullException(nameof(container));
-            }
+            ArgumentNullException.ThrowIfNull(container);
 
-            if (queryText is null)
-            {
-                throw new ArgumentNullException(nameof(queryText));
-            }
+            ArgumentNullException.ThrowIfNull(queryText);
 
-            if (actionAsync is null)
-            {
-                throw new ArgumentNullException(nameof(actionAsync));
-            }
+            ArgumentNullException.ThrowIfNull(actionAsync);
 
             return ForEachAsync(container, new QueryDefinition(queryText), actionAsync, requestOptions, maxBatchCount, continuationToken, cancellationToken);
         }
@@ -165,20 +138,11 @@ namespace Corvus.CosmosClient.Extensions
         /// <returns>A <see cref="Task"/> which provides a continuation token if it terminates before .</returns>
         public static async Task<string?> ForEachAsync<T>(this Container container, QueryDefinition queryDefinition, Func<T, Task> actionAsync, QueryRequestOptions? requestOptions = null, int? maxBatchCount = null, string? continuationToken = null, CancellationToken cancellationToken = default)
         {
-            if (container is null)
-            {
-                throw new ArgumentNullException(nameof(container));
-            }
+            ArgumentNullException.ThrowIfNull(container);
 
-            if (queryDefinition == null)
-            {
-                throw new ArgumentNullException(nameof(queryDefinition));
-            }
+            ArgumentNullException.ThrowIfNull(queryDefinition);
 
-            if (actionAsync == null)
-            {
-                throw new ArgumentNullException(nameof(actionAsync));
-            }
+            ArgumentNullException.ThrowIfNull(actionAsync);
 
             FeedIterator<T> iterator = container.GetItemQueryIterator<T>(queryDefinition, continuationToken, requestOptions);
 
